+++ src/SecucardConnect/Product/Smart/Model/Transaction.php
@@ -24,8 +24,8 @@ class Transaction extends BaseModel
     const STATUS_APPROVED = 'approved';
     const STATUS_OK = 'ok';
 
-    const ORDER_OPTION_COLLECTION = 'collection';
-    const ORDER_OPTION_SHIPPING = 'shipping';
+    const DELIVERY_OPTIONS_COLLECTION = 'collection';
+    const DELIVERY_OPTIONS_SHIPPING = 'shipping';
 
     const CHECKOUT_LAST_VISITED_PAGE_CHECKIN = 'checkin_page';
     const CHECKOUT_LAST_VISITED_PAGE_ADDRESS = 'address_page';
@@ -123,17 +123,7 @@ class Transaction extends BaseModel
     /**
      * @var string
      */
-    public $order_option;
-
-    /**
-     * @var string
-     */
     public $last_visited_page;
-
-    /**
-     * @var PickupOptions
-     */
-    public $pickup_options;
 
     /**
      * @var string
@@ -191,6 +181,11 @@ class Transaction extends BaseModel
     public $payment_links;
 
     /**
+     * @var BaseDeliveryOptions
+     */
+    public $delivery_options;
+
+    /**
      * @return array
      */
     public function jsonFilterNullProperties()
@@ -204,5 +199,27 @@ class Transaction extends BaseModel
             'receipt_merchant_print',
             'error'
         ];
+    }
+
+    /**
+     * @param array $payload
+     * @return $this
+     */
+    public function setDeliveryOptions(array $payload)
+    {
+        if (!isset($payload['type'])) {
+            return $this;
+        }
+
+        switch ($payload['type']) {
+            case self::DELIVERY_OPTIONS_SHIPPING:
+                $this->delivery_options = new DeliveryOptionsShipping();
+                break;
+            case self::DELIVERY_OPTIONS_COLLECTION:
+                $this->delivery_options = new DeliveryOptionsCollection();
+                break;
+        }
+
+        return $this;
     }
 }